# frontend/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/projects/{projectId}/team-members": {
      "put": {
        "tags": [
          "project-controller"
        ],
        "operationId": "addTeamMember",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "teamMemberId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "project-controller"
        ],
        "operationId": "deleteTeamMember",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "teamMemberId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/comments/{commentId}": {
      "put": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "commentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/projects": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "getAll_2",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "project-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/comments/task/{taskId}": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getAll_3",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "create_2",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "text",
            "in": "query",
            "required": true,
            "schema": {
              "maxLength": 500,
              "minLength": 0,
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/registration": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistrationRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "authentication-controller"
        ],
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/attachments/task/{taskId}": {
      "get": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "getAll_4",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "create_3",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "filePath",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tasks/project/{projectId}": {
      "get": {
        "tags": [
          "task-controller"
        ],
        "operationId": "getAll_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/confirm-email": {
      "get": {
        "tags": [
          "authentication-controller"
        ],
        "operationId": "confirmEmail",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}": {
      "delete": {
        "tags": [
          "project-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/projects/{projectId}/quit": {
      "delete": {
        "tags": [
          "project-controller"
        ],
        "operationId": "quitProject",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/attachments/{attachmentId}": {
      "delete": {
        "tags": [
          "attachment-controller"
        ],
        "operationId": "delete_3",
        "parameters": [
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "manager": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        }
      },
      "TaskResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "COMPLETED"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH"
            ]
          },
          "project": {
            "$ref": "#/components/schemas/ProjectResponseDto"
          },
          "responsibleUser": {
            "$ref": "#/components/schemas/UserResponseDto"
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          }
        }
      },
      "CommentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "text": {
            "type": "string"
          },
          "authorId": {
            "type": "integer",
            "format": "int64"
          },
          "taskId": {
            "type": "integer",
            "format": "int64"
          },
          "publicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "changed": {
            "type": "boolean"
          }
        }
      },
...
      "CreateProjectRequestDto": {
        "required": [
          "description",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          }
        }
      },
      "UserRegistrationRequestDto": {
        "required": [
          "email",
          "firstname",
          "password",
          "repeatPassword"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 30,
            "minLength": 8,
            "type": "string"
          },
          "repeatPassword": {
            "maxLength": 30,
            "minLength": 8,
            "type": "string"
          },
          "firstname": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "lastname": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "UserLoginRequestDto": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "maxLength": 30,
            "minLength": 8,
            "type": "string"
          }
        }
      },
      "UserLoginResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "AttachmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "taskId": {
            "type": "integer",
            "format": "int64"
          },
          "taskName": {
            "type": "integer",
            "format": "int64"
          },
          "fileName": {
            "type": "string"
          },
          "filePublicLink": {
            "type": "string"
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}