# src/main/java/bohdan/markatov/org/dto/user/UserLoginResponseDto.java
package bohdan.markatov.org.dto.user;

import lombok.Data;

@Data
public class UserLoginResponseDto {
    private String token;

    public UserLoginResponseDto(String token) {
        this.token = token;
    }
}